requirements.txt
Flask==3.0.3
gunicorn==22.0.0
httpx>=0.27.0
compact_json
requests
databricks-sdk
openai
anyio
streamlit